refactor: use Apache Arrow for S3 instead of s3fs
