+++ packages/vaex-core/vaex/arrow/opener.py
@@ -1,9 +1,13 @@
 import vaex.file
 
 class ArrowOpener:
-    @staticmethod
-    def can_open(path, *args, **kwargs):
+    @classmethod
+    def quick_test(cls, path, *args, **kwargs):
         return path.rpartition('.')[2] == 'arrow'
+
+    @classmethod
+    def can_open(cls, path, *args, **kwargs):
+        return cls.quick_test(path, *args, **kwargs)
 
     @staticmethod
     def open(path, *args, **kwargs):
@@ -11,15 +15,15 @@ class ArrowOpener:
         return open(path, *args, **kwargs)
 
 class ParquetOpener:
-    @staticmethod
-    def can_open(path, *args, **kwargs):
+    @classmethod
+    def quick_test(cls, path, *args, **kwargs):
         return path.rpartition('.')[2] == 'parquet'
+
+    @classmethod
+    def can_open(cls, path, *args, **kwargs):
+        return cls.quick_test(path, *args, **kwargs)
 
     @staticmethod
     def open(path, *args, **kwargs):
         from .dataset import open_parquet
         return open_parquet(path, *args, **kwargs)
-
-def register_opener():
-    vaex.file.register(ArrowOpener)
-    vaex.file.register(ParquetOpener)
