+++ tests/gadget_test.py
@@ -5,8 +5,9 @@ import numpy as np
 def test_open():
     path = 'tests/data/gassphere_littleendian.dat'
     path = 'tests/data/galaxy_littleendian.dat'
-    #ds = vaex.file.other.MemoryMappedGadget(path)
-    ds = vaex.open(path)
+    ds = vaex.dataset.open(path)
     assert ds is not None
-    assert not np.isnan(ds.x.min())
-    print(ds.x.minmax())
+    df = vaex.open(path)
+    assert df is not None
+    assert not np.isnan(df.x.min())
+    print(df.x.minmax())
